convert NavBar to functional compoenent. Passed title and links to NavBar as props form app.js.
